refactor core's structure
