+++ lib/core/theme/custom_text_style.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:frontend/core/utils/size_utils.dart';
-import 'package:frontend/theme/theme_helper.dart';
+import 'package:frontend/core/theme/theme_helper.dart';
 
 /// A collection of pre-defined text styles for customizing text appearance,
 /// categorized by different font families and weights.
