+++ lib/core/widgets/custom_text_form_field.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../core/app_export.dart';
+import '/core/app_export.dart';
 
 class CustomTextFormField extends StatefulWidget {
   CustomTextFormField({
@@ -68,7 +68,7 @@ class _CustomTextFormFieldState extends State<CustomTextFormField> {
   @override
   void initState() {
     super.initState();
-    _isPasswordVisible = widget.obscureText!;
+    _isPasswordVisible = !widget.obscureText!;
   }
 
   void initLogin() {
